Refactor web
