+++ .github/workflows/_zrb.yml
@@ -17,7 +17,7 @@ jobs:
   Run-command:
     runs-on: ubuntu-latest
     container:
-      image: stalchmst/zrb:1.0.0a19
+      image: stalchmst/zrb:1.0.0a21
     steps:
       - name: Check out repository code
         uses: actions/checkout@v3
