+++ src/zrb/runner/cli.py
@@ -6,8 +6,8 @@ from zrb.context.any_context import AnyContext
 from zrb.context.shared_context import SharedContext
 from zrb.group.group import Group
 from zrb.runner.common_util import get_run_kwargs
-from zrb.runner.web_app import create_app
-from zrb.runner.web_config import web_config
+from zrb.runner.web_app import create_web_app
+from zrb.runner.web_config.config_factory import web_config
 from zrb.session.session import Session
 from zrb.session_state_logger.default_session_state_logger import (
     default_session_state_logger,
@@ -167,6 +167,6 @@ server_group = cli.add_group(
 async def start_server(_: AnyContext):
     from uvicorn import Config, Server
 
-    app = create_app(cli, web_config, default_session_state_logger)
+    app = create_web_app(cli, web_config, default_session_state_logger)
     server = Server(Config(app=app, host="0.0.0.0", port=WEB_HTTP_PORT, loop="asyncio"))
     await server.serve()
