+++ src/zrb/runner/web_route/logout_page/controller.py
@@ -1,13 +1,13 @@
 import os
 
 from zrb.group.any_group import AnyGroup
-from zrb.runner.web_config import User
-from zrb.runner.web_util import get_html_auth_link
+from zrb.runner.web_schema.user import User
+from zrb.runner.web_util.html import get_html_auth_link
 from zrb.util.file import read_file
 from zrb.util.string.format import fstring_format
 
 
-def show_login_page(user: User, root_group: AnyGroup):
+def show_logout_page(user: User, root_group: AnyGroup):
     from fastapi.responses import HTMLResponse
 
     _DIR = os.path.dirname(__file__)
@@ -20,6 +20,7 @@ def show_login_page(user: User, root_group: AnyGroup):
                 "name": root_group.name,
                 "description": root_group.description,
                 "auth_link": auth_link,
+                "user": user,
             },
         )
     )
