+++ src/zrb/runner/web_route/static/resources/favicon-32x32.png
