+++ src/zrb/runner/web_route/static/resources/login/event.js
