+++ src/zrb/runner/web_route/static/resources/logout/event.js
