+++ src/zrb/runner/web_route/static/resources/session/current-session.js
