+++ src/zrb/runner/web_route/static/resources/session/event.js
