+++ src/zrb/runner/web_schema/session.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class NewSessionResponse(BaseModel):
+    session_name: str
