+++ src/zrb/runner/web_schema/token.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+
+
+class RefreshTokenRequest(BaseModel):
+    refresh_token: str
+
+
+class Token(BaseModel):
+    access_token: str
+    refresh_token: str
+    token_type: str
